# backend/microservices/auth/auth/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Clinly Auth Service
  version: "1.0.0"
  description: Authentication microservice for the Clinly platform.

servers:
  - url: https://api.clinyl.com
    description: Production server
  - url: http://localhost:8080
    description: Local development

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      summary: Authenticate a user
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    PostRegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: User's password
          example: P@ssw0rd!

    PostLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          example: P@ssw0rd!

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token (24h expiration)
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyQGV4YW1wbGUuY29tIiwicm9sZSI6IlVTRVIiLCJpYXQiOjE3MDAwMDAwMDAsImV4cCI6MTcwMDA4NjQwMH0.signature

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Email already in use
        status:
          type: integer
          description: HTTP status code
          example: 409